Moved fetching data to main app and not landing page
